Kademlia response reuse (#735)
use response from replication-request for replication-response
