+++ modules/command/dh/dh-offer-handle-command.js
@@ -24,7 +24,7 @@ class DHOfferHandleCommand extends Command {
         } = command.data;
 
         this.logger.trace(`Sending replication request for offer ${offerId} to ${dcNodeId}.`);
-        await this.transport.replicationRequest({
+        const response = await this.transport.replicationRequest({
             offerId,
             wallet: this.config.node_wallet,
             dhIdentity: this.config.erc725Identity,
@@ -36,8 +36,45 @@ class DHOfferHandleCommand extends Command {
         bid.status = 'SENT';
         await bid.save({ fields: ['status'] });
 
-        this.logger.notify(`Replication request for ${offerId} sent to ${dcNodeId}`);
-        return Command.empty();
+        this.logger.notify(`Replication request for ${offerId} sent to ${dcNodeId}. Response received.`);
+
+        const packedResponse = DHOfferHandleCommand._stripResponse(response);
+        Object.assign(packedResponse, {
+            dcNodeId,
+        });
+        return {
+            commands: [
+                {
+                    name: 'dhReplicationImportCommand',
+                    data: packedResponse,
+                    transactional: false,
+                },
+            ],
+        };
+    }
+
+    /**
+     * Parse network response
+     * @param response  - Network response
+     * @private
+     */
+    static _stripResponse(response) {
+        return {
+            offerId: response.offer_id,
+            dataSetId: response.data_set_id,
+            edges: response.edges,
+            litigationVertices: response.litigation_vertices,
+            dcWallet: response.dc_wallet,
+            litigationPublicKey: response.litigation_public_key,
+            distributionPublicKey: response.distribution_public_key,
+            distributionPrivateKey: response.distribution_private_key,
+            distributionEpkChecksum: response.distribution_epk_checksum,
+            litigationRootHash: response.litigation_root_hash,
+            distributionRootHash: response.distribution_root_hash,
+            distributionEpk: response.distribution_epk,
+            distributionSignature: response.distribution_signature,
+            transactionHash: response.transaction_hash,
+        };
     }
 
     /**
