+++ modules/service/dh-service.js
@@ -274,73 +274,6 @@ class DHService {
     }
 
     /**
-     * Handle one replication payload
-     * @param offerId   - Offer ID
-     * @param dataSetId - Data set ID
-     * @param litigationVertices  - Encrypted import vertices
-     * @param edges     - Import edges
-     * @param dcWallet  - DC wallet
-     * @param litigationPublicKey - Litigation decryption key
-     * @param distributionPublicKey - Distribution decryption key
-     * @param distributionPrivateKey - Distribution encryption key
-     * @param distributionEpkChecksum - Distribution decryption key checksum
-     * @param dcNodeId  - DC node ID
-     * @param litigationRootHash - Litigation hash
-     * @param distributionRootHash - Distribution hash
-     * @param transactionHash - Transaction hash of the import
-     * @param distributionEpk - Distribution EPK
-     * @param distributionSignature - Distribution parameters check
-     * @returns {Promise<void>}
-     */
-    async handleReplicationImport(
-        offerId,
-        dataSetId,
-        dcNodeId,
-        dcWallet,
-        edges,
-        litigationVertices,
-        litigationPublicKey,
-        distributionPublicKey,
-        distributionPrivateKey,
-        distributionEpkChecksum,
-        litigationRootHash,
-        distributionRootHash,
-        distributionEpk,
-        distributionSignature,
-        transactionHash,
-    ) {
-        const bid = await Models.bids.findOne({
-            where: { offer_id: offerId },
-        });
-        if (!bid) {
-            this.logger.trace(`Request for replication import came for unknown offer ${offerId}. Ignoring.`);
-            return;
-        }
-
-        await this.commandExecutor.add({
-            name: 'dhReplicationImportCommand',
-            data: {
-                offerId,
-                dataSetId,
-                dcNodeId,
-                dcWallet,
-                edges,
-                litigationVertices,
-                litigationPublicKey,
-                distributionPublicKey,
-                distributionPrivateKey,
-                distributionEpkChecksum,
-                litigationRootHash,
-                distributionRootHash,
-                distributionEpk,
-                distributionSignature,
-                transactionHash,
-            },
-            transactional: false,
-        });
-    }
-
-    /**
      * Handle one read request (checks whether node satisfies query)
      * @param msgId       - Message ID
      * @param msgNodeId   - Message node ID
